Refactor Is & Method lowering to separate passes
