+++ tools/src/main/scala/scala/scalanative/compiler/pass/EmptyPass.scala
@@ -2,7 +2,5 @@ package scala.scalanative
 package compiler
 package pass
 
-/**
- * A pass that does nothing.
- */
+/** A pass that does nothing. */
 object EmptyPass extends Pass
