+++ tools/src/main/scala/scala/scalanative/compiler/pass/SizeofLowering.scala
@@ -6,9 +6,7 @@ import scala.collection.mutable
 import util.ScopedVar, ScopedVar.scoped
 import nir._
 
-/** Eliminates:
- *  - Op.SizeOf
- */
+/** Maps sizeof computation to pointer arithmetics over null pointer. */
 class SizeofLowering(implicit fresh: Fresh) extends Pass {
   override def preInst = {
     case Inst(n, Op.Sizeof(ty)) =>
